# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Middleware REST API Doc",
    "description": "Rest API",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "name": "CONUN Blockchain Team"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:4000"
    }
  ],
...
  "paths": {
...
    "/api/v1/users": {
      "post": {
        "tags": ["Login"],
        "summary": "Register",
        "operationId": "register",
        "requestBody": {
          "description": "Register",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/sign"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {}
          },
          "400": {
            "description": "Bad Request",
            "content": {}
          },
          "401": {
            "description": "UnAuthorized",
            "content": {}
          },
          "404": {
            "description": "Not found",
            "content": {}
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/users/me": {
      "get": {
        "tags": ["Login"],
        "summary": "Get current user",
        "operationId": "get me",
        "responses": {
          "200": {
            "description": "Success",
            "content": {}
          },
          "401": {
            "description": "UnAuthorized",
            "content": {}
          },
          "404": {
            "description": "Not found",
            "content": {}
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "sign": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "orgName": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "wallet_address": {
            "type": "string"
          },
          "isAdmin": {
            "type": "boolean"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}